requirements.txt
PySide6==6.5.2
Pillow==10.1.0
